Using int32_t instead of int/long type to solve 64-bits porting issue. Also fixed call of usage() function at command line.

+++ simulations/get_nfw_concentration/rand.c
@@ -14,9 +14,9 @@
  * the unlikely value MASK) to initialize the sequence; Adapted from Numerical 
  * Recipies section 7.1 */
 
-float ran0(long *seed)
+float ran0(int32_t *seed)
 {
-	long k;
+	int32_t k;
 	float ans;
 
 	*seed ^= MASK;
